Remove vendoring of dependencies
Currently the only external dependencies of this project are
go-sql-driver/mysql driver and boltdb. Both of these packages are very
stable so vendoring them is not that important.

We were also vendoring our shimmie package but since we control that
code, it is better to always use the latest version and make sure it
works on all the projects that use it.

Additionally, after creating a new vendor folder through the dep init
command (github.com/golang/dep) we discovered that the resulting vendor
folder will add 6.0M to the repo (for just 3 dependencies).

Vendoring in this case is inconvenient while the simplicity of 'go get'
seems to get the job done just fine. Thus this project will continue
without vendoring for now.